% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{chat_openai}
\alias{chat_openai}
\title{Reimplementation of ellmer::chat_openai to support system messages with
o1-mini and ignore unsupported parameters (e.g. temperature) in all o-series
models}
\usage{
chat_openai(
  system_prompt = NULL,
  turns = NULL,
  base_url = "https://api.openai.com/v1",
  api_key = get_api_key("openai"),
  model = NULL,
  seed = NULL,
  api_args = list(),
  echo = FALSE
)
}
\description{
Reimplementation of ellmer::chat_openai to support system messages with
o1-mini and ignore unsupported parameters (e.g. temperature) in all o-series
models
}
