% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{compile_results_data}
\alias{compile_results_data}
\title{Load results data from temp files}
\usage{
compile_results_data(
  dir = here::here("results", "processed"),
  enrich = TRUE,
  models_path = here::here("inst", "models.csv")
)
}
\arguments{
\item{dir}{Directory containing temp files.}

\item{enrich}{Whether to enrich the results with model configuration}

\item{models_path}{Path to the models CSV file used for enrichment.}
}
\value{
If as = "data.frame", returns a tibble containing all results
combined. If as = "list", returns a list of tibbles, one per file.
}
\description{
Load results data from temp files
}
