% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cthllm-package.R
\docType{package}
\name{cthllm-package}
\alias{cthllm}
\alias{cthllm-package}
\title{Clinical Travel Health LLM Evaluation}
\description{
A package to evaluate Large Language Models (LLMs) performance on clinical
travel health questions. It provides a framework to test different LLMs with
various answering modalities and assess their performance.
}
\author{
\strong{Maintainer}: Angelo D'Ambrosio \email{angelo.dambrosio@ecdc.europa.eu}

Other contributors:
\itemize{
  \item Francesco Baglivo \email{francesco.baglivo@phd.unipi.it} [contributor]
}

}
\keyword{internal}
