% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_env_var}
\alias{get_env_var}
\title{Get environment variable from package root .env file}
\usage{
get_env_var(var_name)
}
\arguments{
\item{var_name}{The name of the environment variable to retrieve.}
}
\value{
The value of the environment variable, or NULL if not found.
}
\description{
Get environment variable from package root .env file
}
\keyword{internal}
