% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_config}
\alias{get_model_config}
\title{Get model configuration from models.csv}
\usage{
get_model_config(model_id)
}
\arguments{
\item{model_id}{The model ID to look up.}
}
\value{
A list containing the model configuration.
}
\description{
Get model configuration from models.csv
}
