% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_models}
\alias{load_models}
\title{Load model configurations}
\usage{
load_models(path = here::here("inst", "models.csv"))
}
\arguments{
\item{path}{Path to the CSV file containing model configurations.}
}
\value{
A tibble containing the model configurations.
}
\description{
Load model configurations
}
