% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_questions}
\alias{load_questions}
\title{Load questions from CSV file}
\usage{
load_questions(path = here::here("inst", "benchmark_travel_medicine.csv"))
}
\arguments{
\item{path}{Path to the CSV file.}
}
\value{
A tibble containing the questions data.
}
\description{
Load questions from CSV file
}
