% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{store_temp_result}
\alias{store_temp_result}
\title{Store a single result in a temp file}
\usage{
store_temp_result(
  result_combination,
  status = NULL,
  dir = here::here("results", "temp")
)
}
\arguments{
\item{result_combination}{A data.frame containing a single result row.}

\item{status}{Result status (C = correct, F = fail, N = not parsable, E =
error).}

\item{dir}{Directory to store temp files.}
}
\value{
The temp file path, invisibly.
}
\description{
Store a single result in a temp file
}
