% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{validate_models}
\alias{validate_models}
\title{Validate model configurations}
\usage{
validate_models(models)
}
\arguments{
\item{models}{Tibble containing model configurations.}
}
\value{
The validated model configurations tibble.
}
\description{
Validate model configurations
}
\keyword{internal}
