% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{validate_questions}
\alias{validate_questions}
\title{Validate questions data}
\usage{
validate_questions(questions)
}
\arguments{
\item{questions}{Tibble containing questions data.}
}
\value{
The validated questions tibble.
}
\description{
Validate questions data
}
\keyword{internal}
