% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{validate_results}
\alias{validate_results}
\title{Validate results data}
\usage{
validate_results(results)
}
\arguments{
\item{results}{Tibble containing results data.}
}
\value{
The validated results tibble.
}
\description{
Validate results data
}
\keyword{internal}
